installation optimization by skipping installed packages

+++ .bin/distro-setup/packages/apt-install.sh
@@ -19,13 +19,26 @@ function ask_install {
     fi
 }
 
-# Prints whether or not package $2 is installed by testing if command $1 exists
+# Returns 1 if package $2 is installed by testing if command $1 exists, 0 otherwise
+function requires_install {
+    if [ -x "$(command -v $1)" ]; then
+        echo -e "\e[1m\e[32m$2 is already installed\e[0m"
+        return 1
+    else
+        echo -e "\e[1m\e[34mInstalling $2...\e[0m"
+        return 0
+    fi   
+}
+
+# Returns 0 if package $2 is installed by testing if command $1 exists, 1 otherwise
 function test_install {
     if [ -x "$(command -v $1)" ]; then
         echo -e "\e[1m\e[32m$2 was successfully installed\e[0m"
+        return 0
     else
         echo -e "\e[1m\e[31m$2 failed to install\e[0m"
-    fi
+        return 1
+    fi   
 }
 
 DEB_CODENAME=$(lsb_release --codename --short)
@@ -40,55 +53,60 @@ apt-get install `tr '\r\n' ' ' < ${SCRIPT_DIR}/packages.txt` -y
 apt-get install wget apt-transport-https ca-certificates curl software-properties-common gnupg-agent python3-pip -y
 
 # Install Dropbox
-{
+cmd="dropbox"
+pkg="Dropbox"
+if requires_install "$cmd" "$pkg" ; then
     wget -O dropbox.deb "https://www.dropbox.com/download?dl=packages/ubuntu/dropbox_2020.03.04_amd64.deb"
     apt-get install "./dropbox.deb" -y --allow-downgrades
     rm dropbox.deb
-}
-test_install "dropbox" "Dropbox"
+
+    test_install "$cmd" "$pkg"
+fi
 
 # Install Visual Studio Code
 # https://linuxize.com/post/how-to-install-visual-studio-code-on-ubuntu-18-04/
-{
+cmd="code"
+pkg="Visual Studio Code"
+if requires_install "$cmd" "$pkg" ; then
     wget -q https://packages.microsoft.com/keys/microsoft.asc -O- | apt-key add -
     add-apt-repository "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main"
     apt-get update -y && apt-get install code -y
-}
-test_install "code" "Visual Studio Code"
+
+    test_install "$cmd" "$pkg"
+fi
 
 # Install VirtualBox
 # https://linuxize.com/post/how-to-install-virtualbox-on-debian-10/
-{
+cmd="virtualbox"
+pkg="VirtualBox"
+if requires_install "$cmd" "$pkg" ; then
     wget -q https://www.virtualbox.org/download/oracle_vbox_2016.asc -O- | apt-key add -
     wget -q https://www.virtualbox.org/download/oracle_vbox.asc -O- | apt-key add -
     add-apt-repository "deb [arch=amd64] http://download.virtualbox.org/virtualbox/debian $(lsb_release -cs) contrib"
     apt-get update -y && apt-get install virtualbox-6.0 -y
-}
-test_install "virtualbox" "VirtualBox"
+
+    test_install "$cmd" "$pkg"
+fi
 
 # Install Docker
 # https://docs.docker.com/engine/install/debian/
-{
+cmd="docker"
+pkg="Docker"
+if requires_install "$cmd" "$pkg" ; then
     apt-get remove docker docker-engine docker.io containerd runc
     curl -fsSL https://download.docker.com/linux/debian/gpg | apt-key add -
     add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/debian $(lsb_release -cs) stable"
     apt-get update -y && apt-get install docker-ce docker-ce-cli containerd.io -y
     usermod -aG docker $USER
-}
-test_install "docker" "Docker"
 
-# # Install Spotify
-# # https://www.spotify.com/us/download/linux/
-# {
-#     curl -sS https://download.spotify.com/debian/pubkey.gpg | apt-key add -
-#     echo "deb http://repository.spotify.com stable non-free" | tee /etc/apt/sources.list.d/spotify.list
-#     apt-get update
-# }
-# test_install "spotify" "Spotify"
+    test_install "$cmd" "$pkg"
+fi
 
 # Build i3 gaps from source
 # https://github.com/Airblader/i3/wiki/Building-from-source
-{
+cmd="i3"
+pkg="i3 gaps"
+if requires_install "$cmd" "$pkg" ; then
     # install dependencies
     apt-get install meson dh-autoreconf libxcb-keysyms1-dev libpango1.0-dev libxcb-util0-dev xcb libxcb1-dev libxcb-icccm4-dev libyajl-dev libev-dev libxcb-xkb-dev libxcb-cursor-dev libxkbcommon-dev libxcb-xinerama0-dev libxkbcommon-x11-dev libstartup-notification0-dev libxcb-randr0-dev libxcb-xrm0 libxcb-xrm-dev libxcb-shape0 libxcb-shape0-dev -y
     
@@ -107,15 +125,19 @@ test_install "docker" "Docker"
     apt-get install i3lock -y
 
     cd $CWD
-}
-test_install "i3" "i3 gaps"
+
+    test_install "$cmd" "$pkg"
+fi
 
 # Install polybar
 # https://github.com/polybar/polybar#installation
-{
+cmd="polybar"
+pkg="polybar"
+if requires_install "$cmd" "$pkg" ; then
     apt -t "$DEB_CODENAME"-backports install polybar -y
-}
-test_install "polybar" "polybar"
+
+    test_install "$cmd" "$pkg"
+fi
 
 # Package installations below saved for posterity:
 
@@ -134,3 +156,12 @@ test_install "polybar" "polybar"
 #     echo "deb https://download.sublimetext.com/ apt/stable/" | sudo tee /etc/apt/sources.list.d/sublime-text.list
 #     apt-get update -y && apt-get install sublime-text -y
 # }
+
+# # Install Spotify
+# # https://www.spotify.com/us/download/linux/
+# {
+#     curl -sS https://download.spotify.com/debian/pubkey.gpg | apt-key add -
+#     echo "deb http://repository.spotify.com stable non-free" | tee /etc/apt/sources.list.d/spotify.list
+#     apt-get update
+# }
+# test_install "spotify" "Spotify"
